:bug: reply, review 엔드포인트 토큰인증 조건문 주석해제

+++ src/controller/reply.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import Reviews_Replies from "../model/Reviews_Replies";
+import { isAuthorized } from "./auth";
 
 // ? 리뷰에 달린 댓글리스트 읽기
 export const reviewList = async (req: Request, res: Response) => {
@@ -18,10 +19,10 @@ export const reviewList = async (req: Request, res: Response) => {
 
 // ? 특정 사용자가 작성한 모든 댓글리스트 읽기.
 export const userList = async (req: Request, res: Response) => {
-  // if (!isAuthorized(req)) {
-  //   res.status(403).json({message: "Invalid Access Token"});
-  //   return;
-  // }
+  if (!isAuthorized(req)) {
+    res.status(403).json({ message: "Invalid Access Token" });
+    return;
+  }
   try {
     const data = await Reviews_Replies.findAll({
       where: { Users_id: req.params.userid },
@@ -37,10 +38,10 @@ export const userList = async (req: Request, res: Response) => {
 
 // ? 리뷰에 새로운 댓글을 생성
 export const create = async (req: Request, res: Response) => {
-  // if (!isAuthorized(req)) {
-  //   res.status(403).json({message: "Invalid Access Token"});
-  //   return;
-  // }
+  if (!isAuthorized(req)) {
+    res.status(403).json({ message: "Invalid Access Token" });
+    return;
+  }
   try {
     const { comment, userId, reviewId } = req.body;
     const data = await Reviews_Replies.create({
@@ -59,10 +60,10 @@ export const create = async (req: Request, res: Response) => {
 
 // ? 리뷰에 있는 댓글 삭제.
 export const deleteReply = async (req: Request, res: Response) => {
-  // if (!isAuthorized(req)) {
-  //   res.status(403).json({message: "Invalid Access Token"});
-  //   return;
-  // }
+  if (!isAuthorized(req)) {
+    res.status(403).json({ message: "Invalid Access Token" });
+    return;
+  }
   try {
     const { replyId } = req.body;
     const data = await Reviews_Replies.destroy({
@@ -79,10 +80,10 @@ export const deleteReply = async (req: Request, res: Response) => {
 
 // ? 리뷰에 있는 댓글 수정
 export const update = async (req: Request, res: Response) => {
-  // if (!isAuthorized(req)) {
-  //   res.status(403).json({message: "Invalid Access Token"});
-  //   return;
-  // }
+  if (!isAuthorized(req)) {
+    res.status(403).json({ message: "Invalid Access Token" });
+    return;
+  }
   try {
     const { comment, replyId } = req.body;
     const data = await Reviews_Replies.update(
